추가 : 글 수정 기능 + 삭제 기능 추가
BbsDAO update, delete 기능 추가
update.jsp, updateAction, delte.jsp 추가함.

+++ StudyWeb/WebContent/view.jsp
@@ -122,7 +122,9 @@
 				if(userID != null && userID.equals(bbs.getUserID())){
 				%>
 			<a href="update.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">수정</a>
-			<a href="delete.jsp?bbsID=<%= bbsID %>" class="btn btn-primary">삭제</a>
+
+			<a onclick="return confirm('정말로 삭제하시겠습니까?')" href="delete.jsp?bbsID=<%=bbsID %>" class="btn btn-primary">삭제</a>
+
 			<%					
 					}
 				%>
